third_party/WebKit/LayoutTests/inspector/sources/pretty-print-javascript-6.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script src="../../http/tests/inspector/sources-test.js"></script>
<script src="debugger/resources/obfuscated.js"></script>

<script>

function test()
{
    var testJSFormatter = InspectorTest.testPrettyPrint.bind(InspectorTest, "text/javascript");

    InspectorTest.runTestSuite([
        function trailingCommentsTest(next)
        {
            var mappingQueries = [];
            testJSFormatter("noop(); //#sourceMappingURL=program.js.map", mappingQueries, next);
        },

        function inlinedScriptFormatting(next)
        {
            var content = "<html><body><script>function f(){}<" + "/script><script>function g(){var a;window.return = 10;if (a) return;}<" + "/script></body></html>";
            InspectorTest.testPrettyPrint("text/html", content, [], next);
        }
    ]);
}

</script>

</head>

<body onload="runTest()">
<p>Verifies JavaScript pretty-printing functionality.</p>
</body>
</html>
