third_party/WebKit/LayoutTests/inspector/sources/pretty-print-javascript-6-expected.txt

Verifies JavaScript pretty-printing functionality.


Running: trailingCommentsTest
====== 8< ------
noop(); //#sourceMappingURL=program.js.map

------ >8 ======

Running: inlinedScriptFormatting
====== 8< ------
<html><body><script>
function f() {
}
</script><script>
function g() {
    var a;
    window.return = 10;
    if (a)
        return;
}
</script></body></html>
------ >8 ======
